% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_NApattern.R
\docType{data}
\name{example_NApattern}
\alias{example_NApattern}
\title{Realistic missing value pattern of proteomic LFQ intensity samples}
\format{A matrix of zero and ones with 1264 rows and 18 columns; 0 means MV, 1 means no MV.}
\source{
\url{https://www.ebi.ac.uk/pride/archive/projects/PXD001584}
}
\usage{
example_NApattern
}
\description{
An exemplary matrix of a missing value (MV) pattern extracted from
measured LFQ intensities of the proteinGroups.txt dataset from PXD001584 [1].
}
\references{
[1] Ramond E, et al. Importance of host cell arginine uptake in Francisella phagosomal escape and ribosomal protein amounts. Mol Cell Proteomics. 2015 14(4):870-881\cr
}
\seealso{
Other data: \code{\link{mbqnBoxplot}},
  \code{\link{mbqnDemo}}, \code{\link{mbqnExample}},
  \code{\link{mbqnSimuData}},
  \code{\link{mbqnSimuDistortion}}
}
\author{
Ariane Schad
}
\concept{data}
\keyword{datasets}
