% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPXDfile.R
\name{getPXDfile}
\alias{getPXDfile}
\title{Download data from PRIDE}
\usage{
getPXDfile(pxd_id, source.path = NULL, file.pattern = "proteingroups")
}
\arguments{
\item{pxd_id}{the PRIDE identifier.}

\item{source.path}{pathname where to store and search for the
data file, e.g. "proteinGroups.txt"-file; default = NULL uses current working
directory.}

\item{file.pattern}{character specifying the kind of dataset for download,
e.g. "proteinGroups" or "peptides".}
}
\value{
status (0=ok, 1=not MaxQuant data set, 2=no proteinGroup file)
}
\description{
Auxiliary function which downloads selected data, like protein
group intensities, from PRIDE [1].
}
\details{
This function requires the R packages rpx [2] and BiocFileChace.
As a temporary
fix old versions of the rpx package's pxurl and pxfiles function were
included
}
\examples{
## Download protein LFQ intensities of PXD001584 from PRIDE
getPXDfile(pxd_id = "PXD001584")
}
\references{
[1] Vizcaíno JA, Csordas A, del-Toro N, Dianes JA, Griss J, Lavidas I,
Mayer G, Perez-Riverol Y, Reisinger F, Ternent T, Xu QW, Wang R, Hermjakob H.
2016 update of the PRIDE database and related tools. Nucleic Acids Res.
2016 Jan 1;44(D1): D447-D456. PubMed PMID:26527722.\cr
[2] See "rpx" by Laurent Gatto (2017). rpx: R Interface to the
ProteomeXchange Repository. Package version 1.10.2,
https://github.com/lgatto/rpx.
}
\author{
Ariane Schad
}
