% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.boxplot_dev.R
\name{mbqn.boxplot_dev}
\alias{mbqn.boxplot_dev}
\title{Boxplot of data matrix where selected row intensities are highlighted}
\usage{
mbqn.boxplot_dev(mtx, irow = NULL, vals = NULL, xlab = NULL,
  ylab = NULL, main = NULL, filename = NULL, type = "l")
}
\arguments{
\item{mtx}{A data matrix. Rows represent features, e.g. protein abundances; columns represent groups, samples, e.g., experimental conditions, replicates.}

\item{irow}{row index or array of row indices for highlighting}

\item{vals}{numeric array of length \code{dim(x)[2]} to plot on top of boxplot}

\item{xlab}{xaxis-label}

\item{ylab}{yaxis-label}

\item{main}{figure title}

\item{filename}{save plot to file with filename}

\item{type}{line style for plot of row intensities}
}
\value{
Figure
}
\description{
Boxplot of a data matrix dominate and highlight selected row indices.
}
\details{
Create boxplot of data and highlight values of selected rows with lines or points
across samples
}
\examples{
qn.dat <- mbqn(x=dat,FUN = NULL ,na.rm = TRUE)
mbqn.dat <- mbqn(x=dat,FUN = median ,na.rm = TRUE)
mbqn.boxplot(qn.dat,irow = 1, vals = mbqn.dat[1,], filename = "fig_boxplot_qn.data.pdf")
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced quantile
normalization for processing label-free quantitative proteomics
data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\keyword{normalization}
\keyword{proteomics}
\keyword{quantile}
