% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.example1.R
\name{mbqn.example1}
\alias{mbqn.example1}
\title{Example illustrating the analysis of PRIDE data with the MBQN package}
\usage{
mbqn.example1(which.example = 1, source.path = NULL)
}
\arguments{
\item{which.example}{Numerical values between 1-4, default = 1,
to select between 4 dataset.}

\item{filepath}{Character with path where to store and search for proteinGroups.txt
file; default = NULL uses "installationpath/MBQN/examples".}
}
\description{
This function runs an example to illustrate the check of LFQ intensities
of proteins for rank invariant (RI) and nearly rank invariant (NRI) features
and the normalization of the data. It downloads a selected dataset from PRIDE
or it uses the dataset in filepath. The example needs the R package "rpx" to
download data from PRIDE.
}
\details{
Collecting information on the experiments requires the package rpx by Laurent Gatto (2017),
version 1.10.2 from https://github.com/lgatto/rpx.\cr
The function used to read the proteinGroups.txt files uses source code
of SafeQuant::parseMaxQuantProteinGroupTxt, \cr
version 2.3.1, https://CRAN.R-project.org/package=SafeQuant.\cr\cr
Experiments on PRIDE that are supported by this Example:\cr
1 PXD001584 - contains a RI (default) (file size 10.8 MB) [1]\cr
2 PXD005138 - contains a RI feature (file size 7440.6 MB) [2]\cr
3 PXD005861 - contains RI feature (file size 334.7 MB) [3]\cr
4 PXD006617 - contains RI feature (file size 290.8 MB) [4]\cr\cr
See the Reference for further details on these examples.
}
\examples{
## Check LFQ intensities of proteomics data PXD001584 downloaded
from PRIDE for RI and NRI features
\dontrun{
example1(which.example = 3)
}
}
\references{
[1] Ramond E, et al. Importance of host cell arginine uptake in Francisella phagosomal escape and ribosomal protein amounts. Mol Cell Proteomics. 2015 14(4):870-881\cr
[2] Beaumont V, et al. Phosphodiesterase 10A Inhibition Improves Cortico-Basal Ganglia Function in Huntington's Disease Models. Neuron. 2016 92(6):1220-1237\cr
[3] Turetschek R, Desalegn G, Epple T, Kaul HP, Wienkoop S. Key metabolic traits of Pisum sativum maintain cell vitality during Didymella pinodes infection: cultivar resistance and the microsymbionts' influence. J Proteomics. 2017 Mar 4. pii: S1874-3919(17)30075-1\cr
[4] Ranjbar Sistani N, Kaul HP, Desalegn G, Wienkoop S. Rhizobium Impacts on Seed Productivity, Quality, and Protection of Pisum sativum upon Disease Stress Caused by Didymella pinodes: Phenotypic, Proteomic, and Metabolomic Traits. Front Plant Sci. 2017 8:1961\cr
[5] Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. Bioinf. Appl. Note, 2018\cr
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
2017
}
\concept{quantile, quantile normalization, rank invariance}
