% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.nri.R
\name{mbqn.nri}
\alias{mbqn.nri}
\title{Selective mean-balanced quantile normalization for (nearly) rank invariant features}
\usage{
mbqn.nri(x, FUN = NULL, na.rm = TRUE, method = NULL, low_thr = 0.5,
  index = NULL)
}
\arguments{
\item{x}{A data matrix, where rows represent proteins and
columns samples from different replicates, treatments, or conditions.}

\item{FUN}{A function like mean, median, or a user defined function, or a numeric array with
\code{dim(user_array) = nrow(x)} to balance each intensity profile across samples.
Functions can be parsed also as characters. If FUN = NULL, features are not balanced, i.e. normal QN is used.}

\item{na.rm}{A logical value indicating whether NA values should be omitted in the computation of average feature expression.}

\item{method}{Function used to compute quantile normalization; default NULL - use function from the preprocessCore package ; if "limma" - the function from the Limma package is used.}

\item{low_thr}{Value between [0 1] as lower threshold that specifies the critical rank invariance frequency.}

\item{index}{A single index or a vector of indices of selected rows.}
}
\value{
Normalized \code{matrix}.
}
\description{
Apply quantile normalization to a data matrix. Selected rows or nearly rank
invariant values with a rank invariance frequency above a threshold are normalized with
the mean balanced quantile normalization.
}
\details{
Quantile normalize a data matrix. Selected rows of the data matrix are normalized with the mean balanced quantile normalization.
}
\examples{
## Apply quantile normalization to a data matrix where nearly rank invariant (NRI) features are balanced
X <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3)
mbqn.nri(X, median,low_thr = 0.5) # Balance NRI features selected by threshold
mbqn.nri(X, median, index = c(1,2)) # Balance selected features
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. Bioinf. Appl. Note, 2018
}
\seealso{
Other mbqn: \code{\link{get_kminmax}},
  \code{\link{mbqn.boxplot}},
  \code{\link{mbqn.check_saturation}}, \code{\link{mbqn}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{mbqn}
\concept{quantile, quantile normalization, rank invariance}
