% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.R
\name{mbqn}
\alias{mbqn}
\title{Mean-balanced quantile normalization}
\usage{
mbqn(x, FUN = NULL, na.rm = TRUE, method = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A data matrix, where rows represent proteins and
columns samples from different replicates, treatments, or conditions.}

\item{FUN}{A function like mean, median, a user defined function, or a numeric array
of weights with
\code{dim(user_array) = nrow(x)} to balance each intensity profile across samples.
Functions can be parsed also as characters. If FUN = NULL, features are not balanced,
i.e. normal QN is used.}

\item{na.rm}{A logical value indicating whether NA values should be omitted in the
computation of average feature expression.}

\item{method}{Packagename providing the function used to compute the standard quantile normalization;
default NULL or "limma" - use \code{normalizeBetweenArrays()} from the limma package,
if "preprocessCore" - \code{normalize.quantiles()} from the preprocessCore package is used.}

\item{verbose}{Logical indicating to run function quietly.}
}
\value{
Normalized matrix
}
\description{
Modified quantile-normalization of a matrix, representing for example
omics or other data sorted in a matrix. Prevents systematic flattening of feature variation across columns
for features overrepresented in the tails of the intensity distribution
across columns, i.e. rank invariant (RI) or nearly rank invariant (NRI) features.
}
\details{
Normalize a data matrix based on a mean-balanced quantile normalization.
Each row of the data matrix is balanced by FUN, e.g. the median, before normalization.
After normalization, row means are added to the normalized matrix.
By default, this function uses \code{limma::normalizeQuantiles()}.
available from http://bioconductor.org/biocLite.R.
}
\examples{
\dontrun{
## Compute mean and median balanced quantile normalization
X <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3)
mbqn(X, mean) # Use arithmetic mean to center features
mbqn(X, median) # Use median to center features
mbqn(X, "median")

## Use user defined array of weights for averaging
wt <- c(1,3,1)/5 # Weights for each sample
user_array <- apply(X,1,weighted.mean, wt ,na.rm =TRUE)
mbqn(X, user_array)
}
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. In prep., 2019
}
\seealso{
Other mbqn: \code{\link{get_kminmax}},
  \code{\link{mbqnCheckSaturation}},
  \code{\link{mbqnGetNRIfeatures}}, \code{\link{mbqnNRI}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{mbqn}
