% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.simu.data.R
\name{mbqn.simu.data}
\alias{mbqn.simu.data}
\title{Generate a random/structured data matrix}
\usage{
mbqn.simu.data(model = "rand", nrow = NULL, ncol = NULL,
  seed = 1234)
}
\arguments{
\item{model}{Two different type of matrix models are available: "rand" generates a random matrix
of size nrow x ncol, "omics" generates a Gaussian random matrix which mimics intensity profiles and
missing values as present in a selected real data set.}

\item{nrow}{number of rows of data matrix.}

\item{ncol}{number of columns of data matrix.}

\item{seed}{Seed for random number generator. Default \code{seed <- 1234}.}
}
\value{
\code{matrix} of size nrow x ncol
}
\description{
Generate a random data matrix with or without proteomics-like properties
}
\details{
For model "rand" the random matrix entries are drawn from a standard normal distribution
of mean 0 and sigma 1. For model "omics" the meand and standard deviation of each row i is
drawn from Gaussian distribution \eqn{N(\mu_i,sd_i^2)} with \eqn{sd_i~N(0,0.0625)} and \eqn{\mu_i~N(28,4)}.
About 21\% of the values are omitted according to the missing value pattern present in the LFQ intensities
of PXD001584 (Ramond et al. 2015).
}
\examples{
mbqn.simu.data(model = "rand", 1000,10)
mbqn.simu.data(model = "rand")
mbqn.simu.data(model = "omics")
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. In prep. 2019. \cr
Ramond, E. et al. (2015) Importance of host cell arginine uptake in Francisella phagosomal
escape and ribosomal protein amounts. Mol Cell Proteomics 14, 870-881.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
