% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnSimuDistortion.R
\name{mbqnSimuDistortion}
\alias{mbqnSimuDistortion}
\title{Simulate distortion of mean and scale of a data matrix}
\usage{
mbqnSimuDistortion(ncol = NULL, max_mean = 0.05, max_scale = 0.01,
  seed = 1234)
}
\arguments{
\item{ncol}{number of columns of data matrix}

\item{max_mean}{Absolute maximum shift of mean from 0 to max_mean}

\item{max_scale}{Realtive change in scale, i.e. 0.01 corresponds to 1\%}

\item{seed}{Seed for random number generator. Default \code{seed <- 1234}}
}
\value{
\code{matrix} of size nrow x ncol
}
\description{
Simulate an omics-like, columnwise distortion in mean and scale
of a data matrix
}
\details{
For model "rand" the random matrix entries are drawn from a standard normal distribution
of mean 0 and sigma 1. For model "omics" the meand and standard deviation of each row i is
drawn from Gaussian distribution \eqn{N(\mu_i,sd_i^2)} with \eqn{sd_i~N(0,0.0625)} and \eqn{\mu_i~N(28,4)}.
About 21\% of the values are omitted according to the missing value pattern present in the LFQ intensities
of PXD001584 (Ramond et al. 2015).
}
\examples{
\dontrun{
MBQN:::mbqnSimuDistortion(ncol = 9)
}
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. In prep. 2019. \cr
Ramond, E. et al. (2015) Importance of host cell arginine uptake in Francisella phagosomal
escape and ribosomal protein amounts. Mol Cell Proteomics 14, 870-881.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
