% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnThreshold.R
\name{oneSidedTest}
\alias{oneSidedTest}
\title{Recalulate p value from two-sided to one-sided}
\usage{
oneSidedTest(sign_value, z_value)
}
\arguments{
\item{sign_value}{P value from two-sided significance test}

\item{z_value}{Z value from two-sided significance test}
}
\value{
P value from one sided significance test
}
\description{
Recalulate p value from two-sided to one-sided
}
