% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnThreshold.R
\name{truncate_decimals}
\alias{truncate_decimals}
\title{Truncate float to defined number of decimal values}
\usage{
truncate_decimals(x, digits = 2)
}
\arguments{
\item{x}{float}

\item{digits}{Number of decimal values}
}
\value{
Truncated number
}
\description{
Truncate float to defined number of decimal values
}
