% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnThreshold.R
\name{truncateDecimals}
\alias{truncateDecimals}
\title{Truncate float to defined number of decimal values}
\usage{
truncateDecimals(x, digits = 2)
}
\arguments{
\item{x}{float}

\item{digits}{Number of decimal values}
}
\value{
Truncated number
}
\description{
Truncate float to defined number of decimal values
}
\examples{
x <- 2.567836
truncateDecimals(x, 3)
}
