% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kminmax.R
\name{get_kminmax}
\alias{get_kminmax}
\title{Get the k largest/smallest values and their indices for each
column of a matrix}
\usage{
get_kminmax(X, k, flag)
}
\arguments{
\item{X}{data matrix.}

\item{k}{number of selected extreme values. Must be less then #rows of X.}

\item{flag}{"min or "max"}
}
\value{
\code{list} with indices \code{ik} of extreme values \code{minmax}.
}
\description{
Extract the k largest/smallest values and their indices for each column of a matrix.
}
\details{
Function used by \code{mbqn.check_saturation}. Sort values of
each column of a matrix or an array X and determine the
k smallest (flag <- "min") or largest (flag <- "max") values in. NA's in the
data are ignored.
}
\examples{
X <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3) # Create a data matrix
get_kminmax(X, k = 5, "max") # get indices of the 5 largest values in each column
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. Bioinf. Appl. Note., 2018
}
\seealso{
Other mbqn: \code{\link{mbqn.boxplot}},
  \code{\link{mbqn.check_saturation}},
  \code{\link{mbqn.nri}}, \code{\link{mbqn}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{mbqn}
\concept{quantile, quantile normalization, rank invariance}
