% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kminmax.R
\name{get_kminmax}
\alias{get_kminmax}
\title{Get the k largest/smallest elements}
\usage{
get_kminmax(x, k, flag = "max")
}
\arguments{
\item{x}{a data matrix or data frame.}

\item{k}{an integer specifying the number of extreme values. Must be \code{<= nrows(x)}.}

\item{flag}{use "min" or "max" (default) to select smallest or largest elements.}
}
\value{
List with elements:
\item{\code{ik}}{indices of ordered extreme values}
\item{\code{minmax}}{ordered extreme values.}
}
\description{
Extract the k largest or smallest values and their indices for each
column of a matrix.
}
\details{
Order the values of each column of \code{x} and determine the
k smallest (\code{flag = "min"}) or largest (\code{flag = "max"}) values and their indices.
NA's in the data are ignored.
}
\examples{
x <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3) # Create a data matrix
MBQN:::get_kminmax(x, k = 5, "max") # get indices of the 5 largest values in each column
}
\references{
Schad, A. and Kreutz, C., MBQN: R package for mean/median-balanced quantile
normalization. In prep. 2019
}
\author{
Ariane Schad
}
