% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kminmax.R
\name{get_kminmax}
\alias{get_kminmax}
\title{Get the k largest/smallest values and their indices for each
column of a matrix}
\usage{
get_kminmax(X, k, flag)
}
\arguments{
\item{X}{data matrix.}

\item{k}{number of selected extreme values. Must be less then #rows of X.}

\item{flag}{"min or "max"}
}
\value{
\code{list} with indices \code{ik} of extreme values \code{minmax}.
}
\description{
Extract the k largest/smallest values and their indices for each column of a matrix.
}
\details{
Function used by \code{mbqn.check_saturation}. Sort values of
each column of a matrix or an array X and determine the
k smallest (flag <- "min") or largest (flag <- "max") values in. NA's in the
data are ignored.
}
\examples{
get_kminmax(X, k = 5, "max")
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced
quantile normalization for processing label-free quantitative
proteomics data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\keyword{normalization,}
\keyword{proteomics}
\keyword{quantile}
