% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{loadFile}
\alias{loadFile}
\title{Load and preprocess LFQ intensities}
\usage{
loadFile(pxd_id, source.path = NULL, file.pattern = "proteinGroups")
}
\arguments{
\item{pxd_id}{character specifying the PRIDE identifier.}

\item{source.path}{character specifying the path where to store and search for the
"proteinGroups.txt"-file; default = NULL uses current working directory.}

\item{file.pattern}{character specifying the kind of dataset for download, e.g.
"proteinGroups.txt" or "peptides.txt".}
}
\value{
List with
\item{\code{mtx}}{data matrix}
\item{\code{pxdid}}{PRIDE identifier}
\item{\code{featureAnnotations}}{dataframe collecting feature annotations, e.g. protein name,
"Potential contaminant", etc..}
}
\description{
Auxilary function used to load and preprocess LFQ
intensities of selected dataset with a PRIDE [1] identifier.
}
\details{
Load proteinGroup.txt file, select all samples with LFQ intensities, remove
empty protein features. Apply log2 transform to intensities. This function acquires source code
of SafeQuant::parseMaxQuantProteinGroupTxt [2].
}
\examples{
## Load LFQ intensities of proteomics data PXD001584 downloaded
available at PRIDE:
\dontrun{
loadFile(pxd_id = "PXD001584")
}
}
\references{
[1] Vizcaíno JA, Csordas A, del-Toro N, Dianes JA, Griss J, Lavidas I, Mayer G,
Perez-Riverol Y, Reisinger F, Ternent T, Xu QW, Wang R, Hermjakob H.
2016 update of the PRIDE database and related tools. Nucleic Acids Res.
2016 Jan 1;44(D1): D447-D456. PubMed PMID:26527722.\cr
[2] See "SafeQuant" by Erik Ahrne (2016). SafeQuant: A Toolbox for the Analysis of Proteomics Data. R package version 2.3.1.,
https://CRAN.R-project.org/package=SafeQuant.\cr
}
\seealso{
Other example: \code{\link{mbqnBoxplot}},
  \code{\link{mbqnDemo}}, \code{\link{mbqnExample}}
}
\author{
Ariane Schad
}
\concept{example}
