% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.check_saturation.R
\name{mbqn.check_saturation}
\alias{mbqn.check_saturation}
\title{Check data matrix for rank invariant (ri) /nearly rank invariant (nri) features}
\usage{
mbqn.check_saturation(dat, FUN = NULL, flag_show_fig = TRUE,
  low_thr = 0.2, filename = NULL, feature_index = NULL, method = NULL)
}
\arguments{
\item{dat}{A data matrix. Rows - features, e.g. protein abundances; columns - samples}

\item{FUN}{median or mean, if left empty, quantile normalization
is applied without balancing the data}

\item{flag_show_fig}{flag to specify whether results are plotted to figure, default = TRUE}

\item{low_thr}{lower threshold for NRI frequency, default = 0.2}

\item{filename}{string for naming figures, default = NULL}

\item{feature_index}{index of a feature of interest that is plotted in the boxplot; default NULL}

\item{method}{Package used to compute quantile normalization; default NULL - use the preprocessCore package ; "limma" uses the Limma package}
}
\value{
A matrix of median- or mean-balanced quantile normalized data
}
\description{
Test if few rows of the data matrix dominate the upper tail. This script uses normalize.quantiles() from the
package preprocessCore that can be installed from http://bioconductor.org/biocLite.R
}
\details{
Rank data and check if lower and upper intensity tails are
dominated by few feature. Compute a quantile
normalization without and with mean-balancing and check standard
deviation of normalized data entries located in the tails
}
\examples{
mbqn.check_saturation(dat, mean)
}
\references{
Schad, A. and Kreuz, C. (2017) Median-balanced quantile
normalization for processing label-free quantitative proteomics
data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
2017
}
\keyword{normalization}
\keyword{proteomics}
\keyword{quantile}
