% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.check_saturation.R
\name{mbqn.check_saturation}
\alias{mbqn.check_saturation}
\title{Test if normalization of data benefits from mean-balancing}
\usage{
mbqn.check_saturation(dat, FUN = mean_fun, qlow, qup)
}
\arguments{
\item{dat}{A data matrix. Rows - features, e.g. protein abundances; columns - samples}

\item{qlow}{lower quantile}

\item{qup}{upper quantile, default 1}

\item{mean_fun}{mean or median, if left empty, quantile normalization is applied without balancing the data}
}
\value{
A matrix of mean- or median-balanced quantile normalized data
}
\description{
This function uses normalize.quantiles() from the package preprocessCore that can be installed from http://bioconductor.org/biocLite.R
}
\details{
Compute a quantile normalization without and with mean-balancing and check standard deviation of normalized data entries located in the tails
}
\examples{
mbqn(dat, mean)
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced quantile normalization for processing label-free quantitative proteomics data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, Aug. 2017
}
\keyword{normalization}
\keyword{proteomics}
\keyword{quantile}
