% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.demo.R
\name{mbqn.demo}
\alias{mbqn.demo}
\title{Demonstration of mean/median-balanced quantile normalization}
\usage{
mbqn.demo(dat = NULL)
}
\arguments{
\item{dat}{A data matrix where rows represent features, e.g. protein
abundances/intensities, and columns experimental samples or
(technical/biological) replicates.
Default NULL - a simple matrix is generated.}
}
\value{
Various graphics illustrating saturation in case of
quantile normalization of high expression for a single protein
abundances across samples
}
\description{
This function demonstrates mean-balanced quantile normalization
}
\details{
This function uses \code{normalize.quantiles()} from the package
preprocessCore that can be installed from http://bioconductor.org/biocLite.R by
source('http://bioconductor.org/biocLite.R')
biocLite('preprocessCore'). A data matrix can be computed with mbqn.simu_dat().
}
\examples{
mbqn.demo
mbqn.demo(dat)
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced quantile
normalization for processing label-free quantitative proteomics
data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
Aug. 2017
}
\keyword{normalization}
\keyword{proteomics}
\keyword{quantile}
