% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.R
\name{mbqn}
\alias{mbqn}
\title{Mean-balanced quantile normalization}
\usage{
mbqn(x, FUN = NULL, na.rm = TRUE, method = NULL)
}
\arguments{
\item{x}{A matrix where rows represent features, e.g. protein abundances/intensities and
columns are replicates or probes.}

\item{FUN}{A function like mean, median, or a user defined function or an array with
\code{dim(user_array) = nrow(x)}. Default NULL - features are not balanced.}

\item{method}{Function to compute quantile normalization; default NULL - use function from the preprocessCore package ; if "limma" - the function from the Limma package is used.}
}
\value{
Mean-/Median-balanced quantile normalized \code{matrix}.
}
\description{
Modified quantile-normalization of a matrix, representing for example
omics or microarray data. Suppresses systematic flattening of feature variation across columns
for features overrepresented in the tails of the intensity distribution
across columns, i.e. rank invariant (RI) or nearly rank invariant (NRI) features.
}
\details{
Normalize a data matrix based on a mean-balanced quantile normalization.
Each row of the data matrix is balanced by FUN, e.g. the median, before normalization.
After normalization, row means are added to the normalized matrix.
This function uses \code{preprocessCore::normalize.quantiles()} by Bolstad et al, Bioinformatics (2003),
available from http://bioconductor.org/biocLite.R.
}
\examples{
mbqn(x, median)
mbqn(x, mean)
mbqn(x, user_function)
mbqn(x, user_array)
mbqn(x, median, method = "limma") # if Limma R package is installed and preferred
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. Bioinf. Appl. Note., 2018
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\keyword{Modified}
\keyword{Quantile}
\keyword{normalization,}
\keyword{proteomics.}
