% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.R
\name{mbqn}
\alias{mbqn}
\title{Mean-balanced quantile normalization}
\usage{
mbqn(x, FUN = NULL, na.rm = TRUE, method = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A data matrix, where rows represent proteins and
columns samples from different replicates, treatments, or conditions.}

\item{FUN}{A function like mean, median, a user defined function, or a numeric array
of weights with
\code{dim(user_array) = nrow(x)} to balance each intensity profile across samples.
Functions can be parsed also as characters. If FUN = NULL, features are not balanced,
i.e. normal QN is used.}

\item{na.rm}{A logical value indicating whether NA values should be omitted in the
computation of average feature expression.}

\item{method}{Method used to compute quantile normalization; default NULL - use QN
function from the preprocessCore package ; if "limma" - the QN function from the
Limma package is used.}

\item{verbose}{Logical indicating to run function quietly}
}
\value{
Mean-/Median-balanced quantile normalized \code{matrix}.
}
\description{
Modified quantile-normalization of a matrix, representing for example
omics or other data sorted in a matrix. Prevents systematic flattening of feature variation across columns
for features overrepresented in the tails of the intensity distribution
across columns, i.e. rank invariant (RI) or nearly rank invariant (NRI) features.
}
\details{
Normalize a data matrix based on a mean-balanced quantile normalization.
Each row of the data matrix is balanced by FUN, e.g. the median, before normalization.
After normalization, row means are added to the normalized matrix.
This function uses \code{preprocessCore::normalize.quantiles()} by Bolstad et al, Bioinformatics (2003),
available from http://bioconductor.org/biocLite.R.
}
\examples{
## Compute mean balanced quantile normalization
X <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3)
mbqn(X, mean) # Use arithmetic mean to center features
mbqn(X, median) # Use median to center features
mbqn(X, "median")

\dontrun{
## Use user defined array of weights for averaging
wt <- c(1,3,1)/5 # Weights for each sample
user_array <- apply(X,1,weighted.mean, wt ,na.rm =TRUE)
mbqn(X, user_array)

## Use limma package to compute quantile normalization
mbqn(X, median, method = "limma")}
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. In prep., 2019
}
\seealso{
Other mbqn: \code{\link{get_kminmax}},
  \code{\link{mbqn.boxplot}},
  \code{\link{mbqn.check_saturation}},
  \code{\link{mbqn.get_nri_features}},
  \code{\link{mbqn.nri}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{mbqn}
\concept{quantile, quantile normalization, rank invariance}
