% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.R
\name{mbqn}
\alias{mbqn}
\title{Mean-balanced quantile normalization}
\usage{
mbqn(dat, FUN = NULL)
}
\arguments{
\item{dat}{A data matrix. Rows - features, e.g. protein abundances; columns - samples}

\item{FUN}{mean or median, or an array with dim = numbers of rows of dat. If left empty, quantile normalization is applied without balancing the data}
}
\value{
A matrix of mean- or median-balanced quantile normalized data
}
\description{
This function uses normalize.quantiles() from the package preprocessCore that can be installed from http://bioconductor.org/biocLite.R
}
\details{
Function to compute a mean-balanced quantile normalization
}
\examples{
mbqn(dat, mean)
mbqn(dat,median)
mbqn(dat, x)
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced quantile normalization for processing label-free quantitative proteomics data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, Aug. 2017
}
\keyword{normalization}
\keyword{proteomics}
\keyword{quantile}
