% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqn.R
\name{mbqn}
\alias{mbqn}
\title{Mean-balanced quantile normalization}
\usage{
mbqn(x, FUN = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A matrix where rows represent features, e.g. protein abundances/intensities and
columns are replicates or probes.}

\item{FUN}{A function like mean or median or a user defined function or an array with
\code{dim(user_array) = nrow(x)}. If left empty, the matrix is not balanced.}
}
\value{
Mean-/Median-balanced quantile normalized \code{matrix}.
}
\description{
Modified quantile-normalization of a matrix representing
omics or microarray data. Suppress systematic flattening of feature variation across columns
for features overrepresented in the tails of the intensity distribution
across columns.
}
\details{
Function to normalize a data matrix based on a mean-balanced quantile normalization.
Each row of the data matrix is balanced by its mean/median before normalization.
Row means are added to the normalized matrix. This function uses the preprocessCor
 R package by Bolstad et al, Bioinformatics (2003). This function uses \code{normalize.quantiles()} from the package
preprocessCore that can be installed from http://bioconductor.org/biocLite.R by
source('http://bioconductor.org/biocLite.R') biocLite('preprocessCore').
}
\examples{
mbqn(x, mean)
mbqn(x, median)
mbqn(x, user_function)
mbqn(x, user_array)
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced quantile
normalization for processing label-free quantitative proteomics
data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\keyword{Modified}
\keyword{Quantile}
\keyword{normalization}
\keyword{proteomics.}
