% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnCheckSaturation.R
\name{mbqnCheckSaturation}
\alias{mbqnCheckSaturation}
\title{Check data matrix for rank invariant (RI) and nearly rank invariant (NRI) features}
\usage{
mbqnCheckSaturation(dat, FUN = NULL, low_thr = 0.5,
  feature_index = NULL, show_fig = TRUE, save_fig = TRUE,
  show_nri_only = FALSE, filename = NULL, verbose = TRUE, ...)
}
\arguments{
\item{dat}{A data matrix. Rows - features, e.g. protein abundances; columns - samples}

\item{FUN}{median, mean, or another function used to balance features across colums. If left empty, quantile normalization
is applied without balancing the data}

\item{low_thr}{Numerical value for the lower threshold for NRI frequency, default = 0.5}

\item{feature_index}{Integer that indicates the index of a feature of interest that is plotted in the boxplot; default NULL}

\item{show_fig}{Logical flag indicating whether results should be displayed; default = TRUE.}

\item{save_fig}{Logical to save figures to file}

\item{show_nri_only}{Logical to print and save only the RI/NRI detection graph; default FALSE}

\item{filename}{String for naming figures, default = NULL}

\item{verbose}{Logical for running function quiet}

\item{...}{Optional plot arguments passed to \code{mbqn.boxplot}}
}
\value{
List with the components:
\item{\code{p}}{a matrix of the rank invariance frequencies and the sample coverage over all RI/NRI features}
\item{\code{max_p}}{value of the maximum rank invariance frequency in percent}
\item{\code{ip}}{index of the feature with maximum rank invariance frequency}
\item{\code{nri}}{a table of the rank invariance frequencies in percent of NRI/RI features}
\item{\code{var0_feature}}{index of features with zero sample variance after QN.}
}
\description{
Check data matrix for intensity features in rows which dominate the upper tail, i.e. for
features that have constant or nearly constant rank across samples.
}
\details{
Rank data and check if lower and upper intensity tails are
dominated by few features. Apply quantile
normalization without and with mean-balancing and check the standard
deviation of normalized features located in the tails.
}
\examples{
## Check data matrix for RI and NRI features
X <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3)
mbqnCheckSaturation(X, mean, low_thr = 0.5, save_fig = FALSE)
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. Bioinf. Appl. Note., 2018
}
\seealso{
Other mbqn: \code{\link{get_kminmax}},
  \code{\link{mbqnGetNRIfeatures}}, \code{\link{mbqnNRI}},
  \code{\link{mbqn}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{mbqn}
\concept{quantile, quantile normalization, rank invariance}
