% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnDemo.R
\name{mbqnDemo}
\alias{mbqnDemo}
\title{Demonstration of mean/median-balanced quantile normalization}
\usage{
mbqnDemo(dat = NULL)
}
\arguments{
\item{dat}{a matrix where rows represent features, e.g. protein
abundances/intensities, and columns represent experimental samples.
Default \code{NULL} - a simple matrix is generated.}
}
\value{
Various graphics illustrating saturation in case of
quantile normalization of high expression for a single protein
abundances across samples
}
\description{
This function demonstrates mean/median-balanced quantile normalization
}
\details{
Normalize a matrix and return boxplots of quantile normalized and mean balanced normalized
data. An omics-like data matrix can be generated with \code{mbqn.simu.dat()}.
}
\examples{
mbqnDemo()
dat <- mbqnSimuData()
mbqnDemo(dat)
}
\references{
Schad, A. and Kreuz, C. (2017) Mean-balanced quantile
normalization for processing label-free quantitative proteomics
data with abundance-isolated proteins. Biostatistics xxx in prep.
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\keyword{normalization}
\keyword{proteomics}
\keyword{quantile}
