% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnDemo.R
\name{mbqnDemo}
\alias{mbqnDemo}
\title{Demonstrate mean/median-balanced quantile normalization}
\usage{
mbqnDemo(x = NULL)
}
\arguments{
\item{x}{a matrix where rows represent features, e.g. of protein
abundances/intensities, and columns represent samples. Default
\code{NULL} - a simple matrix is generated.}
}
\value{
Various graphics illustrating the effect of normalization on
rank mixing and rank invariant intensity features.
}
\description{
Visualize the intensity distribution of unnormalized,
quantile and mean/median-balanced quantile normalized data.
}
\details{
Normalize matrix and return boxplots of quantile normalized and mean balanced normalized
data.
}
\examples{
mbqnDemo()
x <- mbqnSimuData()
mbqnDemo(x)
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean/median-balanced quantile normalization. In prep. 2019
}
\seealso{
Other example: \code{\link{mbqnBoxplot}},
  \code{\link{mbqnExample}}
}
\author{
Ariane Schad
}
\concept{example}
