% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnNRI.R
\name{mbqnNRI}
\alias{mbqnNRI}
\title{Selective mean-balanced quantile normalization for (nearly) rank invariant features}
\usage{
mbqnNRI(x, FUN = NULL, na.rm = TRUE, method = NULL, low_thr = 0.5,
  index = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A data matrix, where rows represent proteins and
columns samples from different replicates, treatments, or conditions.
function from the preprocessCore package ; if "limma" - the QN function from the
Limma package is used.
computation of average feature expression.}

\item{FUN}{A function like mean, median, a user defined function, or a numeric array
of weights with
\code{dim(user_array) = nrow(x)} to balance each intensity profile across samples.
Functions can be parsed also as characters. If FUN = NULL, features are not balanced,
i.e. normal QN is used.}

\item{na.rm}{A logical value indicating whether NA values should be omitted in the
computation of average feature expression.}

\item{method}{Method used to compute quantile normalization; default NULL - use QN
function from the preprocessCore package ; if "limma" - the QN function from the
Limma package is used.}

\item{low_thr}{Value between [0 1] as lower threshold that specifies the critical
rank invariance frequency, with default set to 0.5.}

\item{index}{A single index or a vector of indices of selected rows.}

\item{verbose}{Logical indicating to run function quietly}
}
\value{
Normalized \code{matrix}.
}
\description{
Quantile normalization of a data matrix where selected rows or rank invariant (RI)/
nearly rank invariant (NRI) rows/features are normalized by
the mean/median balanced quantile normalization.
}
\details{
Specified rows and/or rows with a rank invariance frequency between samples above \code{threshold}
are normalized with the mean/median balanced quantile normalization. Remaining rows are quantile normalized
without balancing of means.
}
\examples{
## Quantile normalize a data matrix where
## nearly rank invariant (NRI) features are balanced
X <- matrix(c(5,2,3,NA,4,1,4,2,3,4,6,NA,1,3,1),ncol=3)
mbqnNRI(X, median,low_thr = 0.5) # Balance NRI features selected by threshold
mbqnNRI(X, median, index = c(1,2)) # Balance selected features
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean balanced quantile normalization. In prep. 2019
}
\seealso{
Other mbqn: \code{\link{get_kminmax}},
  \code{\link{mbqnBoxplot}},
  \code{\link{mbqnCheckSaturation}},
  \code{\link{mbqnGetNRIfeatures}}, \code{\link{mbqn}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{mbqn}
\concept{quantile, quantile normalization, rank invariance}
