% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnPlotRI.R
\name{mbqnPlotRI}
\alias{mbqnPlotRI}
\title{Plot frequency of detected RI/NRI features}
\usage{
mbqnPlotRI(obj, save_fig = FALSE, filename = NULL, verbose = FALSE,
  ...)
}
\arguments{
\item{obj}{list object of RI frequencies from \code{mbqnGetNRIfeatures()}.}

\item{save_fig}{logical indicating to save figure to pdf.}

\item{filename}{string to specify the filename of the .pdf.}

\item{verbose}{logical indicating to run function quietly.}

\item{...}{additional arguments (cex, cex.lab, cex.axis, cex.main) passed to the plot function,
and (width, height, paper format(default "a4") of pdf) to \code{dev.copy2pdf()}.}
}
\value{
Figure, optionally saved in file Fig_nri_check_" \code{filename} ".pdf.
}
\description{
Plot rank invariance frequency and feature coverage of detected RI and NRI features
}
\details{
Graphical output of the NRI/RI identification results from \code{mbqnGetNRIfeatures()}.
For each detected NRI/RI feature, plot the feature index against the RI frequencies
together with the RI frequency detection threshold and print the sample coverage.
}
\examples{
## Check data matrix for RI and NRI features
\dontrun{
x <- mbqnSimuData("omics.dep")
RI <- mbqnGetNRIfeatures(x, low_thr = 0.5, verbose = FALSE)
mbqnPlotRI(RI)
}
}
\references{
Schad, A. and Kreuz, C., MBQN: R package for mean/median-balanced quantile normalization. In prep. 2019
}
\seealso{
\code{\link[=mbqnGetNRIfeatures]{mbqnGetNRIfeatures()}} for detection of NRI/RI features.
}
\author{
Ariane Schad
}
