% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbqnSimuData.R
\name{mbqnSimuData}
\alias{mbqnSimuData}
\title{Generate a random/structured data matrix}
\usage{
mbqnSimuData(model = "rand", nrow = NULL, ncol = NULL, seed = 1234)
}
\arguments{
\item{model}{Two different type of matrix models are available: "rand" generates a random matrix
of size nrow x ncol, "omics" generates a Gaussian random matrix which mimics intensity profiles and
missing values as present in a real data set.}

\item{nrow}{number of rows of data matrix.}

\item{ncol}{number of columns of data matrix.}

\item{seed}{Seed for random number generator. Default \code{seed <- 1234}.}
}
\value{
\code{matrix} of size nrow x ncol
}
\description{
Generate a random data matrix with or without proteomics, log-transformed feature
intensity-like properties
}
\details{
For model "rand" the matrix entries are drawn from a standard normal
distribution \eqn{N(0,1)}. For model "omics" the matrix entries of each row
are drawn from a Gaussian distribution \eqn{N(\mu_i,\sigma_i^2)} where the
mean and standard deviation itself are drawn Gaussian distributions, i.e.
\eqn{\sigma_i~N(0,0.0625)} and \eqn{\mu_i~N(28,4)}. About 35\% of the matrix
values are set to NA according to the missing value pattern present in the protein LFQ
intensities of PXD001584 [1].
}
\examples{
mbqnSimuData(model = "rand", 1000,10)
mbqnSimuData(model = "rand")
mbqnSimuData(model = "omics")
}
\references{
[1] Ramond, E. et al. (2015) Importance of host cell arginine uptake in Francisella phagosomal
escape and ribosomal protein amounts. Mol Cell Proteomics 14, 870-881.\cr
}
\seealso{
\code{\link[MBQN:example_NApattern]{MBQN::example_NApattern()}} for description of missing value pattern

Other data: \code{\link{example_NApattern}},
  \code{\link{mbqnBoxplot}}, \code{\link{mbqnDemo}},
  \code{\link{mbqnExample}},
  \code{\link{mbqnSimuDistortion}}
}
\author{
A. Schad, \email{ariane.schad@zbsa.de}
}
\concept{data}
